fix TCP Close test
