+++ stack/socket_tcp.go
@@ -1,7 +1,6 @@
 package stack
 
 import (
-	"cmp"
 	"errors"
 	"io"
 	"log/slog"
@@ -140,7 +139,7 @@ func (t *TCPSocket) Close() error {
 
 func (t *TCPSocket) handleMain(response []byte, pkt *TCPPacket) (n int, err error) {
 	defer func() {
-		if err != nil && t.abortErr == nil {
+		if err != nil && t.abortErr == nil && err != io.ErrNoProgress {
 			err = nil // Only close socket if socket is aborted.
 		} else if err != nil {
 			t.stack.error("tcp socket", slog.Int("port", int(t.localPort)), slog.String("err", err.Error()))
@@ -211,7 +210,6 @@ func (t *TCPSocket) handleSend(response []byte, pkt *TCPPacket) (n int, err erro
 	if err != nil {
 		return 0, err
 	}
-	t.setSrcDest(pkt)
 
 	// If we have user data to send we send it, else we send the control segment.
 	var payload []byte
@@ -222,9 +220,14 @@ func (t *TCPSocket) handleSend(response []byte, pkt *TCPPacket) (n int, err erro
 			panic("bug in handleUser") // This is a bug in ring buffer or a race condition.
 		}
 	}
+	t.setSrcDest(pkt)
 	pkt.CalculateHeaders(seg, payload)
 	pkt.PutHeaders(response)
-	return sizeTCPNoOptions + n, nil
+
+	if t.scb.HasPending() {
+		err = io.ErrNoProgress // Flag to PortStack that we have pending data to send.
+	}
+	return sizeTCPNoOptions + n, err
 }
 
 func (t *TCPSocket) setSrcDest(pkt *TCPPacket) {
@@ -274,107 +277,4 @@ func (t *TCPSocket) abort(err error) error {
 	t.abortErr = err
 	t.close()
 	return err
-}
-
-type ring struct {
-	buf []byte
-	off int
-	end int
-}
-
-func (r *ring) Write(b []byte) (int, error) {
-	free := r.Free()
-	if len(b) > free {
-		return 0, errors.New("no more space")
-	}
-	midFree := r.midFree()
-	if midFree > 0 {
-		n := copy(r.buf[r.end:], b)
-		r.end += n
-		return n, nil
-	}
-
-	n := copy(r.buf[r.end:], b)
-	r.end = n
-	if n < len(b) {
-		n2 := copy(r.buf, b[n:])
-		r.end = n2
-		n += n2
-	}
-	return n, nil
-}
-
-func (r *ring) Read(b []byte) (int, error) {
-	if r.Buffered() == 0 {
-		return 0, io.EOF
-	}
-
-	if r.end >= r.off {
-		// start       off       end      len(buf)
-		//   |  sfree   |  used   |  efree   |
-		n := copy(b, r.buf[r.off:r.end])
-		r.off += n
-		r.onReadEnd()
-		return n, nil
-	}
-	// start     end       off     len(buf)
-	//   |  used  |  mfree  |  used  |
-	n := copy(b, r.buf[r.off:])
-	r.off += n
-	if n < len(b) {
-		n2 := copy(b[n:], r.buf[:r.end])
-		r.off = n2
-		n += n2
-	}
-	r.onReadEnd()
-	return n, nil
-}
-
-func (r *ring) Buffered() int {
-	return len(r.buf) - r.Free()
-}
-
-func (r *ring) Reset() {
-	r.off = 0
-	r.end = 0
-}
-
-func (r *ring) Free() int {
-	if r.end >= r.off {
-		// start       off       end      len(buf)
-		//   |  sfree   |  used   |  efree   |
-		startFree := r.off
-		endFree := len(r.buf) - r.end
-		return startFree + endFree
-	}
-	// start     end       off     len(buf)
-	//   |  used  |  mfree  |  used  |
-	return r.off - r.end
-}
-
-func (r *ring) midFree() int {
-	if r.end >= r.off {
-		return 0
-	}
-	return r.off - r.end
-}
-
-func (r *ring) onReadEnd() {
-	if r.off == r.end {
-		r.Reset() // We read everything, reset.
-	}
-}
-
-func max[T cmp.Ordered](a, b T) T {
-	if a > b {
-		return a
-	}
-	return b
-}
-
-func min[T cmp.Ordered](a, b T) T {
-	if a < b {
-		return a
-	}
-	return b
 }
